resolved build errors
