+++ components/issue/style.module.scss
@@ -36,7 +36,7 @@
     align-items: center;
     justify-content: center;
     font-size: 1.1rem;
-    background-color: pink;
+    // background-color: pink;
     display: flex;
 }
 .checkIcon {
