+++ pages/index.js
@@ -1,7 +1,7 @@
 import Head from 'next/head'
 import classes from '../styles/globalStyle.module.scss'
 import Carousel from '../components/carousel/carousel';
-import PersonaCard from '../components/personaCard/personaCard';
+import PersonaCard from '../components/personas/personaCard';
 import Communities from '../components/communities/communities';
 import Events from '../components/events/events';
 import Announcement from '../components/announcement/announcementCards';
